+++ src/custom.js
@@ -8,8 +8,7 @@ var custom_config = {
         "listen_mode": false,
         "shower_timer": false,
         "shower_alert": false,
-        "publish_time": 120,
-        "heating_circuit": 1
+        "publish_time": 120
     }
 }
 
@@ -22,7 +21,6 @@ function listcustom() {
     document.getElementById("led_gpio").value = custom_config.settings.led_gpio;
     document.getElementById("dallas_gpio").value = custom_config.settings.dallas_gpio;
     document.getElementById("publish_time").value = custom_config.settings.publish_time;
-    document.getElementById("heating_circuit").value = custom_config.settings.heating_circuit;
     document.getElementById("tx_mode").value = custom_config.settings.tx_mode;
 
     if (custom_config.settings.led) {
@@ -72,7 +70,6 @@ function savecustom() {
     }
 
     custom_config.settings.publish_time = parseInt(document.getElementById("publish_time").value);
-    custom_config.settings.heating_circuit = parseInt(document.getElementById("heating_circuit").value);
     custom_config.settings.tx_mode = parseInt(document.getElementById("tx_mode").value);
 
     custom_uncommited();
@@ -159,7 +156,7 @@ function listCustomStats() {
 
         document.getElementById("hm").innerHTML = ajaxobj.hp.hm;
         document.getElementById("hp1").innerHTML = ajaxobj.hp.hp1 + " &#37;";
-        document.getElementById("hp2").innerHTML = ajaxobj.hp.hp2 + " &#37;";      
+        document.getElementById("hp2").innerHTML = ajaxobj.hp.hp2 + " &#37;";
     } else {
         document.getElementById("hp_show").style.display = "none";
     }