+++ tools/wsemulator/wserver.js
@@ -105,7 +105,6 @@ var custom_configfile = {
         "shower_timer": true,
         "shower_alert": false,
         "publish_time": 120,
-        "heating_circuit": 1,
         "tx_mode": 1
     }
 };